ai-service/test_groq_connection.py
import os
from dotenv import load_dotenv
from groq import Groq

load_dotenv()
api_key = os.getenv("GROQ_API_KEY")
print(f"API Key found: {api_key[:10]}...")

try:
    client = Groq(api_key=api_key)
    print("Testing connection...")
    response = client.chat.completions.create(
        model="llama-3.3-70b-versatile",
        messages=[{"role": "user", "content": "Say hello"}],
        max_tokens=10,
    )
    print(f"Response: {response.choices[0].message.content}")
except Exception as e:
    print(f"ERROR: {str(e)}")
